Refactors galore!
