+++ README.md
@@ -1,9 +1,13 @@
 # mkdoxin
 
+<p align="center">
+  <img src="https://raw.githubusercontent.com/tritones/mkdoxin/main/.github/assets/mkdoxin-logo.svg" width="320" alt="mkdoxin!">
+</p>
+
+> When life gives you docs, make dachshund!
+
 [![MIT license](https://img.shields.io/badge/License-MIT-blue.svg)](https://github.com/tritones/mkdoxin/blob/main/LICENSE)
-![build status](https://github.com/tritones/mkdoxin/actions/workflows/build.yml/badge.svg?branch=main)
+[![build status](https://github.com/tritones/mkdoxin/actions/workflows/build.yml/badge.svg?branch=main)](https://github.com/tritones/mkdoxin/actions)
 [![GitHub release](https://img.shields.io/github/release/tritones/mkdoxin.svg)](https://github.com/tritones/mkdoxin/releases/)
 
 mkdoxin takes external documentation repositories and builds them using mkdocs (including mkdocs-material).
-
-> When life gives you docs, make dachshund!
