+++ mkdoxin/scheduler.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+from logger import logger
+import schedule
+import time
+
+log = logger(__name__)
