+++ requirements.txt
@@ -2,3 +2,4 @@ giturlparse>=0.10.0
 mkdocs>=1.3.0
 PyYAML>=5.1
 requests>=2.28.1
+schedule>=1.1.0
